requirements.txt
phi
tavily-python
openpyxl
phi-tools
python-dotenv
groq
phidata